feat: メニュー画像生成ハンドラーをmenuItemHandlerに移動し、関連するインポートを修正
